chore: implement base layer for request providers
