+++ packages/core/src/storages/index.ts
@@ -1,6 +1,7 @@
 export * from './dataset';
 export * from './key_value_store';
 export * from './request_list';
+export * from './request_provider';
 export * from './request_queue';
 export * from './request_queue_v2';
 export * from './storage_manager';
