+++ packages/core/src/storages/request_queue.ts
@@ -1,22 +1,11 @@
 import { setTimeout as sleep } from 'node:timers/promises';
 
 import { REQUEST_QUEUE_HEAD_MAX_LIMIT } from '@apify/consts';
-import { ListDictionary, LruCache } from '@apify/datastructures';
-import { cryptoRandomObjectId } from '@apify/utilities';
-import type {
-    BatchAddRequestsResult,
-    Dictionary,
-    ProcessedRequest,
-    QueueOperationInfo,
-    RequestQueueClient,
-    RequestQueueInfo,
-    StorageClient,
+import type { Dictionary,
 } from '@crawlee/types';
-import { chunk, downloadListOfUrls } from '@crawlee/utils';
-import ow from 'ow';
 
-import type { StorageManagerOptions } from './storage_manager';
-import { StorageManager } from './storage_manager';
+import type { RequestProviderOptions } from './request_provider';
+import { RequestProvider } from './request_provider';
 import {
     API_PROCESSED_REQUESTS_DELAY_MILLIS,
     MAX_QUERIES_FOR_CONSISTENCY,
@@ -24,13 +13,9 @@ import {
     QUERY_HEAD_MIN_LENGTH,
     STORAGE_CONSISTENCY_DELAY_MILLIS,
     getRequestId,
-    purgeDefaultStorages,
 } from './utils';
 import { Configuration } from '../configuration';
-import { log } from '../log';
-import type { ProxyConfiguration } from '../proxy_configuration';
-import type { InternalSource, RequestOptions, Source } from '../request';
-import { Request } from '../request';
+import type { Request } from '../request';
 
 const MAX_CACHED_REQUESTS = 1_000_000;
 
@@ -40,36 +25,6 @@ const MAX_CACHED_REQUESTS = 1_000_000;
  * @internal
  */
 const RECENTLY_HANDLED_CACHE_SIZE = 1000;
-
-/**
- * @internal
- */
-interface RequestQueueOperationInfo extends QueueOperationInfo {
-
-    /** Indicates if request was already present in the queue. */
-    wasAlreadyPresent: boolean;
-
-    /** Indicates if request was already marked as handled. */
-    wasAlreadyHandled: boolean;
-
-    /** The ID of the added request */
-    requestId: string;
-
-    uniqueKey: string;
-
-}
-
-export interface RequestQueueOperationOptions {
-    /**
-     * If set to `true`:
-     *   - while adding the request to the queue: the request will be added to the foremost position in the queue.
-     *   - while reclaiming the request: the request will be placed to the beginning of the queue, so that it's returned
-     *   in the next call to {@apilink RequestQueue.fetchNextRequest}.
-     * By default, it's put to the end of the queue.
-     * @default false
-     */
-    forefront?: boolean;
-}
 
 /**
  * Represents a queue of URLs to crawl, which is used for deep crawling of websites
@@ -117,22 +72,8 @@ export interface RequestQueueOperationOptions {
  * ```
  * @category Sources
  */
-export class RequestQueue {
-    log = log.child({ prefix: 'RequestQueue' });
-    id: string;
-    name?: string;
-    timeoutSecs = 30;
-    clientKey = cryptoRandomObjectId();
-    client: RequestQueueClient;
-    private proxyConfiguration?: ProxyConfiguration;
-
-    /**
-     * Contains a cached list of request IDs from the head of the queue,
-     * as obtained in the last query. Both key and value is the request ID.
-     * Need to apply a type here to the generated TS types don't try to use types-apify
-     */
-    private queueHeadDict = new ListDictionary<string>();
-    queryQueueHeadPromise?: Promise<{
+export class RequestQueue extends RequestProvider {
+    private queryQueueHeadPromise?: Promise<{
         wasLimitReached: boolean;
         prevLimit: number;
         queueModifiedAt: Date;
@@ -140,329 +81,16 @@ export class RequestQueue {
         hadMultipleClients?: boolean;
     }> | null = null;
 
-    // A set of all request IDs that are currently being handled,
-    // i.e. which were returned by fetchNextRequest() but not markRequestHandled()
-    inProgress = new Set();
-
-    // To track whether the queue gets stuck, and we need to reset it
-    // `lastActivity` tracks the time when we either added, processed or reclaimed a request,
-    // or when we add new request to in-progress cache
-    lastActivity = new Date();
-    internalTimeoutMillis = 5 * 60e3; // defaults to 5 minutes, will be overridden by BasicCrawler
-
-    // Contains a list of recently handled requests. It is used to avoid inconsistencies
-    // caused by delays in the underlying DynamoDB storage.
-    // Keys are request IDs, values are true.
-    recentlyHandled = new LruCache({ maxLength: RECENTLY_HANDLED_CACHE_SIZE });
-
-    // We can trust these numbers only in a case that queue is used by a single client.
-    // This information is returned by getHead() under the hadMultipleClients property.
-    assumedTotalCount = 0;
-    assumedHandledCount = 0;
-
-    // Caching requests to avoid redundant addRequest() calls.
-    // Key is computed using getRequestId() and value is { id, isHandled }.
-    requestsCache = new LruCache<
-        { uniqueKey: string; wasAlreadyHandled: boolean; isHandled: boolean; id: string }
-    >({ maxLength: MAX_CACHED_REQUESTS });
-
     /**
      * @internal
      */
-    constructor(options: RequestQueueOptions, readonly config = Configuration.getGlobalConfig()) {
-        this.id = options.id;
-        this.name = options.name;
-        this.client = options.client.requestQueue(this.id, {
-            clientKey: this.clientKey,
-            timeoutSecs: this.timeoutSecs,
-        }) as RequestQueueClient;
-        this.proxyConfiguration = options.proxyConfiguration;
-    }
-
-    /**
-     * @ignore
-     */
-    inProgressCount() {
-        return this.inProgress.size;
-    }
-
-    /**
-     * Adds a request to the queue.
-     *
-     * If a request with the same `uniqueKey` property is already present in the queue,
-     * it will not be updated. You can find out whether this happened from the resulting
-     * {@apilink QueueOperationInfo} object.
-     *
-     * To add multiple requests to the queue by extracting links from a webpage,
-     * see the {@apilink enqueueLinks} helper function.
-     *
-     * @param requestLike {@apilink Request} object or vanilla object with request data.
-     * Note that the function sets the `uniqueKey` and `id` fields to the passed Request.
-     * @param [options] Request queue operation options.
-     */
-    async addRequest(requestLike: Source, options: RequestQueueOperationOptions = {}): Promise<RequestQueueOperationInfo> {
-        ow(requestLike, ow.object);
-        ow(options, ow.object.exactShape({
-            forefront: ow.optional.boolean,
-        }));
-
-        this.lastActivity = new Date();
-        const { forefront = false } = options;
-
-        if ('requestsFromUrl' in requestLike) {
-            const requests = await this._fetchRequestsFromUrl(requestLike as InternalSource);
-            const processedRequests = await this._addFetchedRequests(requestLike as InternalSource, requests, options);
-
-            return processedRequests[0];
-        }
-
-        ow(requestLike, ow.object.partialShape({
-            url: ow.string,
-            id: ow.undefined,
-        }));
-
-        const request = requestLike instanceof Request
-            ? requestLike
-            : new Request(requestLike);
-
-        const cacheKey = getRequestId(request.uniqueKey);
-        const cachedInfo = this.requestsCache.get(cacheKey);
-
-        if (cachedInfo) {
-            request.id = cachedInfo.id;
-            return {
-                wasAlreadyPresent: true,
-                // We may assume that if request is in local cache then also the information if the
-                // request was already handled is there because just one client should be using one queue.
-                wasAlreadyHandled: cachedInfo.isHandled,
-                requestId: cachedInfo.id,
-                uniqueKey: cachedInfo.uniqueKey,
-            };
-        }
-
-        const queueOperationInfo = await this.client.addRequest(request, { forefront }) as RequestQueueOperationInfo;
-        queueOperationInfo.uniqueKey = request.uniqueKey;
-
-        const { requestId, wasAlreadyPresent } = queueOperationInfo;
-        this._cacheRequest(cacheKey, queueOperationInfo);
-
-        if (!wasAlreadyPresent && !this.inProgress.has(requestId) && !this.recentlyHandled.get(requestId)) {
-            this.assumedTotalCount++;
-
-            // Performance optimization: add request straight to head if possible
-            this._maybeAddRequestToQueueHead(requestId, forefront);
-        }
-
-        return queueOperationInfo;
-    }
-
-    /**
-     * Adds requests to the queue in batches of 25.
-     *
-     * If a request that is passed in is already present due to its `uniqueKey` property being the same,
-     * it will not be updated. You can find out whether this happened by finding the request in the resulting
-     * {@apilink BatchAddRequestsResult} object.
-     *
-     * @param requestsLike {@apilink Request} objects or vanilla objects with request data.
-     * Note that the function sets the `uniqueKey` and `id` fields to the passed requests if missing.
-     * @param [options] Request queue operation options.
-     */
-    async addRequests(
-        requestsLike: Source[],
-        options: RequestQueueOperationOptions = {},
-    ): Promise<BatchAddRequestsResult> {
-        ow(requestsLike, ow.array);
-        ow(options, ow.object.exactShape({
-            forefront: ow.optional.boolean,
-        }));
-
-        const { forefront = false } = options;
-
-        const uniqueKeyToCacheKey = new Map<string, string>();
-        const getCachedRequestId = (uniqueKey: string) => {
-            const cached = uniqueKeyToCacheKey.get(uniqueKey);
-
-            if (cached) return cached;
-
-            const newCacheKey = getRequestId(uniqueKey);
-            uniqueKeyToCacheKey.set(uniqueKey, newCacheKey);
-
-            return newCacheKey;
-        };
-
-        const results: BatchAddRequestsResult = {
-            processedRequests: [],
-            unprocessedRequests: [],
-        };
-
-        for (const requestLike of requestsLike) {
-            if ('requestsFromUrl' in requestLike) {
-                const requests = await this._fetchRequestsFromUrl(requestLike as InternalSource);
-                await this._addFetchedRequests(requestLike as InternalSource, requests, options);
-            }
-        }
-
-        const requests = requestsLike
-            .filter((requestLike) => !('requestsFromUrl' in requestLike))
-            .map((requestLike) => {
-                return requestLike instanceof Request ? requestLike : new Request(requestLike as RequestOptions);
-            });
-
-        const requestsToAdd = new Map<string, Request>();
-
-        for (const request of requests) {
-            const cacheKey = getCachedRequestId(request.uniqueKey);
-            const cachedInfo = this.requestsCache.get(cacheKey);
-
-            if (cachedInfo) {
-                request.id = cachedInfo.id;
-                results.processedRequests.push({
-                    wasAlreadyPresent: true,
-                    // We may assume that if request is in local cache then also the information if the
-                    // request was already handled is there because just one client should be using one queue.
-                    wasAlreadyHandled: cachedInfo.isHandled,
-                    requestId: cachedInfo.id,
-                    uniqueKey: cachedInfo.uniqueKey,
-                });
-            } else if (!requestsToAdd.has(request.uniqueKey)) {
-                requestsToAdd.set(request.uniqueKey, request);
-            }
-        }
-
-        // Early exit if all provided requests were already added
-        if (!requestsToAdd.size) {
-            return results;
-        }
-
-        const apiResults = await this.client.batchAddRequests([...requestsToAdd.values()], { forefront });
-
-        // Report unprocessed requests
-        results.unprocessedRequests = apiResults.unprocessedRequests;
-
-        // Add all new requests to the queue head
-        for (const newRequest of apiResults.processedRequests) {
-            // Add the new request to the processed list
-            results.processedRequests.push(newRequest);
-
-            const cacheKey = getCachedRequestId(newRequest.uniqueKey);
-
-            const { requestId, wasAlreadyPresent } = newRequest;
-            this._cacheRequest(cacheKey, newRequest);
-
-            if (!wasAlreadyPresent && !this.inProgress.has(requestId) && !this.recentlyHandled.get(requestId)) {
-                this.assumedTotalCount++;
-
-                // Performance optimization: add request straight to head if possible
-                this._maybeAddRequestToQueueHead(requestId, forefront);
-            }
-        }
-
-        return results;
-    }
-
-    /**
-     * Adds requests to the queue in batches. By default, it will resolve after the initial batch is added, and continue
-     * adding the rest in background. You can configure the batch size via `batchSize` option and the sleep time in between
-     * the batches via `waitBetweenBatchesMillis`. If you want to wait for all batches to be added to the queue, you can use
-     * the `waitForAllRequestsToBeAdded` promise you get in the response object.
-     *
-     * @param requests The requests to add
-     * @param options Options for the request queue
-     */
-    async addRequestsBatched(requests: (string | Source)[], options: AddRequestsBatchedOptions = {}): Promise<AddRequestsBatchedResult> {
-        ow(requests, ow.array.ofType(ow.any(
-            ow.string,
-            ow.object.partialShape({ url: ow.string, id: ow.undefined }),
-            ow.object.partialShape({ requestsFromUrl: ow.string, regex: ow.optional.regExp }),
-        )));
-        ow(options, ow.object.exactShape({
-            forefront: ow.optional.boolean,
-            waitForAllRequestsToBeAdded: ow.optional.boolean,
-            batchSize: ow.optional.number,
-            waitBetweenBatchesMillis: ow.optional.number,
-        }));
-
-        const {
-            batchSize = 1000,
-            waitBetweenBatchesMillis = 1000,
-        } = options;
-        const builtRequests: Request[] = [];
-
-        for (const opts of requests) {
-            if (opts && typeof opts === 'object' && 'requestsFromUrl' in opts) {
-                await this.addRequest(opts, { forefront: options.forefront });
-            } else {
-                builtRequests.push(new Request(typeof opts === 'string' ? { url: opts } : opts as RequestOptions));
-            }
-        }
-
-        const attemptToAddToQueueAndAddAnyUnprocessed = async (providedRequests: Request[]) => {
-            const resultsToReturn: ProcessedRequest[] = [];
-            const apiResult = await this.addRequests(providedRequests, { forefront: options.forefront });
-            resultsToReturn.push(...apiResult.processedRequests);
-
-            if (apiResult.unprocessedRequests.length) {
-                await sleep(waitBetweenBatchesMillis);
-
-                resultsToReturn.push(...await attemptToAddToQueueAndAddAnyUnprocessed(
-                    providedRequests.filter((r) => !apiResult.processedRequests.some((pr) => pr.uniqueKey === r.uniqueKey)),
-                ));
-            }
-
-            return resultsToReturn;
-        };
-
-        const initialChunk = builtRequests.splice(0, batchSize);
-
-        // Add initial batch of `batchSize` to process them right away
-        const addedRequests = await attemptToAddToQueueAndAddAnyUnprocessed(initialChunk);
-
-        // If we have no more requests to add, return early
-        if (!builtRequests.length) {
-            return {
-                addedRequests,
-                waitForAllRequestsToBeAdded: Promise.resolve([]),
-            };
-        }
-
-        // eslint-disable-next-line no-async-promise-executor
-        const promise = new Promise<ProcessedRequest[]>(async (resolve) => {
-            const chunks = chunk(builtRequests, batchSize);
-            const finalAddedRequests: ProcessedRequest[] = [];
-
-            for (const requestChunk of chunks) {
-                finalAddedRequests.push(...await attemptToAddToQueueAndAddAnyUnprocessed(requestChunk));
-
-                await sleep(waitBetweenBatchesMillis);
-            }
-
-            resolve(finalAddedRequests);
-        });
-
-        // If the user wants to wait for all the requests to be added, we wait for the promise to resolve for them
-        if (options.waitForAllRequestsToBeAdded) {
-            addedRequests.push(...await promise);
-        }
-
-        return {
-            addedRequests,
-            waitForAllRequestsToBeAdded: promise,
-        };
-    }
-
-    /**
-     * Gets the request from the queue specified by ID.
-     *
-     * @param id ID of the request.
-     * @returns Returns the request object, or `null` if it was not found.
-     */
-    async getRequest<T extends Dictionary = Dictionary>(id: string): Promise<Request<T> | null> {
-        ow(id, ow.string);
-
-        const requestOptions = await this.client.getRequest(id);
-        if (!requestOptions) return null;
-
-        return new Request(requestOptions as unknown as RequestOptions);
+    constructor(options: RequestProviderOptions, config = Configuration.getGlobalConfig()) {
+        super({
+            ...options,
+            logPrefix: 'RequestQueue',
+            recentlyHandledRequestsMaxSize: RECENTLY_HANDLED_CACHE_SIZE,
+            requestCacheMaxSize: MAX_CACHED_REQUESTS,
+        }, config);
     }
 
     /**
@@ -482,25 +110,25 @@ export class RequestQueue {
      * @returns
      *   Returns the request object or `null` if there are no more pending requests.
      */
-    async fetchNextRequest<T extends Dictionary = Dictionary>(): Promise<Request<T> | null> {
+    override async fetchNextRequest<T extends Dictionary = Dictionary>(): Promise<Request<T> | null> {
         await this._ensureHeadIsNonEmpty();
 
-        const nextRequestId = this.queueHeadDict.removeFirst();
+        const nextRequestId = this.queueHeadIds.removeFirst();
 
         // We are likely done at this point.
         if (!nextRequestId) return null;
 
         // This should never happen, but...
-        if (this.inProgress.has(nextRequestId) || this.recentlyHandled.get(nextRequestId)) {
+        if (this.requestIdsInProgress.has(nextRequestId) || this.recentlyHandledRequestsCache.get(nextRequestId)) {
             this.log.warning('Queue head returned a request that is already in progress?!', {
                 nextRequestId,
-                inProgress: this.inProgress.has(nextRequestId),
-                recentlyHandled: !!this.recentlyHandled.get(nextRequestId),
+                inProgress: this.requestIdsInProgress.has(nextRequestId),
+                recentlyHandled: !!this.recentlyHandledRequestsCache.get(nextRequestId),
             });
             return null;
         }
 
-        this.inProgress.add(nextRequestId);
+        this.requestIdsInProgress.add(nextRequestId);
         this.lastActivity = new Date();
 
         let request;
@@ -508,7 +136,7 @@ export class RequestQueue {
             request = await this.getRequest(nextRequestId);
         } catch (e) {
             // On error, remove the request from in progress, otherwise it would be there forever
-            this.inProgress.delete(nextRequestId);
+            this.requestIdsInProgress.delete(nextRequestId);
             throw e;
         }
 
@@ -522,7 +150,7 @@ export class RequestQueue {
         if (!request) {
             this.log.debug('Cannot find a request from the beginning of queue, will be retried later', { nextRequestId });
             setTimeout(() => {
-                this.inProgress.delete(nextRequestId);
+                this.requestIdsInProgress.delete(nextRequestId);
             }, STORAGE_CONSISTENCY_DELAY_MILLIS);
             return null;
         }
@@ -533,146 +161,16 @@ export class RequestQueue {
         //    will not put the request again to queueHeadDict.
         if (request.handledAt) {
             this.log.debug('Request fetched from the beginning of queue was already handled', { nextRequestId });
-            this.recentlyHandled.add(nextRequestId, true);
+            this.recentlyHandledRequestsCache.add(nextRequestId, true);
             return null;
         }
 
         return request;
     }
 
-    /**
-     * Marks a request that was previously returned by the
-     * {@apilink RequestQueue.fetchNextRequest}
-     * function as handled after successful processing.
-     * Handled requests will never again be returned by the `fetchNextRequest` function.
-     */
-    async markRequestHandled(request: Request): Promise<RequestQueueOperationInfo | null> {
-        this.lastActivity = new Date();
-        ow(request, ow.object.partialShape({
-            id: ow.string,
-            uniqueKey: ow.string,
-            handledAt: ow.optional.string,
-        }));
-
-        if (!this.inProgress.has(request.id)) {
-            this.log.debug(`Cannot mark request ${request.id} as handled, because it is not in progress!`, { requestId: request.id });
-            return null;
-        }
-
-        const handledAt = request.handledAt ?? new Date().toISOString();
-        const queueOperationInfo = await this.client.updateRequest({ ...request, handledAt }) as RequestQueueOperationInfo;
-        request.handledAt = handledAt;
-        queueOperationInfo.uniqueKey = request.uniqueKey;
-
-        this.inProgress.delete(request.id);
-        this.recentlyHandled.add(request.id, true);
-
-        if (!queueOperationInfo.wasAlreadyHandled) {
-            this.assumedHandledCount++;
-        }
-
-        this._cacheRequest(getRequestId(request.uniqueKey), queueOperationInfo);
-
-        return queueOperationInfo;
-    }
-
-    /**
-     * Reclaims a failed request back to the queue, so that it can be returned for processing later again
-     * by another call to {@apilink RequestQueue.fetchNextRequest}.
-     * The request record in the queue is updated using the provided `request` parameter.
-     * For example, this lets you store the number of retries or error messages for the request.
-     */
-    async reclaimRequest(request: Request, options: RequestQueueOperationOptions = {}): Promise<RequestQueueOperationInfo | null> {
-        this.lastActivity = new Date();
-        ow(request, ow.object.partialShape({
-            id: ow.string,
-            uniqueKey: ow.string,
-        }));
-        ow(options, ow.object.exactShape({
-            forefront: ow.optional.boolean,
-        }));
-
-        const { forefront = false } = options;
-
-        if (!this.inProgress.has(request.id)) {
-            this.log.debug(`Cannot reclaim request ${request.id}, because it is not in progress!`, { requestId: request.id });
-            return null;
-        }
-
-        // TODO: If request hasn't been changed since the last getRequest(),
-        //   we don't need to call updateRequest() and thus improve performance.
-        const queueOperationInfo = await this.client.updateRequest(request, { forefront }) as RequestQueueOperationInfo;
-        queueOperationInfo.uniqueKey = request.uniqueKey;
-        this._cacheRequest(getRequestId(request.uniqueKey), queueOperationInfo);
-
-        // Wait a little to increase a chance that the next call to fetchNextRequest() will return the request with updated data.
-        // This is to compensate for the limitation of DynamoDB, where writes might not be immediately visible to subsequent reads.
-        setTimeout(() => {
-            if (!this.inProgress.has(request.id)) {
-                this.log.debug('The request is no longer marked as in progress in the queue?!', { requestId: request.id });
-                return;
-            }
-
-            this.inProgress.delete(request.id);
-
-            // Performance optimization: add request straight to head if possible
-            this._maybeAddRequestToQueueHead(request.id, forefront);
-        }, STORAGE_CONSISTENCY_DELAY_MILLIS);
-
-        return queueOperationInfo;
-    }
-
-    /**
-     * Resolves to `true` if the next call to {@apilink RequestQueue.fetchNextRequest}
-     * would return `null`, otherwise it resolves to `false`.
-     * Note that even if the queue is empty, there might be some pending requests currently being processed.
-     * If you need to ensure that there is no activity in the queue, use {@apilink RequestQueue.isFinished}.
-     */
-    async isEmpty(): Promise<boolean> {
+    protected override async ensureHeadIsNonEmpty(): Promise<void> {
+        // Alias for backwards compatibility
         await this._ensureHeadIsNonEmpty();
-        return this.queueHeadDict.length() === 0;
-    }
-
-    /**
-     * Resolves to `true` if all requests were already handled and there are no more left.
-     * Due to the nature of distributed storage used by the queue,
-     * the function might occasionally return a false negative,
-     * but it will never return a false positive.
-     */
-    async isFinished(): Promise<boolean> {
-        if ((Date.now() - +this.lastActivity) > this.internalTimeoutMillis) {
-            const message = `The request queue seems to be stuck for ${this.internalTimeoutMillis / 1e3}s, resetting internal state.`;
-            this.log.warning(message, { inProgress: [...this.inProgress] });
-            this._reset();
-        }
-
-        if (this.queueHeadDict.length() > 0 || this.inProgressCount() > 0) return false;
-
-        const isHeadConsistent = await this._ensureHeadIsNonEmpty(true);
-        return isHeadConsistent && this.queueHeadDict.length() === 0 && this.inProgressCount() === 0;
-    }
-
-    private _reset() {
-        this.queueHeadDict.clear();
-        this.queryQueueHeadPromise = null;
-        this.inProgress.clear();
-        this.recentlyHandled.clear();
-        this.assumedTotalCount = 0;
-        this.assumedHandledCount = 0;
-        this.requestsCache.clear();
-        this.lastActivity = new Date();
-    }
-
-    /**
-     * Caches information about request to beware of unneeded addRequest() calls.
-     */
-    protected _cacheRequest(cacheKey: string, queueOperationInfo: RequestQueueOperationInfo): void {
-        this.requestsCache.add(cacheKey, {
-            id: queueOperationInfo.requestId,
-            isHandled: queueOperationInfo.wasAlreadyHandled,
-            uniqueKey: queueOperationInfo.uniqueKey,
-            wasAlreadyHandled: queueOperationInfo.wasAlreadyHandled,
-        });
     }
 
     /**
@@ -692,7 +190,7 @@ export class RequestQueue {
         iteration = 0,
     ): Promise<boolean> {
         // If is nonempty resolve immediately.
-        if (this.queueHeadDict.length() > 0) return true;
+        if (this.queueHeadIds.length() > 0) return true;
 
         if (!this.queryQueueHeadPromise) {
             const queryStartedAt = new Date();
@@ -702,9 +200,9 @@ export class RequestQueue {
                 .then(({ items, queueModifiedAt, hadMultipleClients }) => {
                     items.forEach(({ id: requestId, uniqueKey }) => {
                         // Queue head index might be behind the main table, so ensure we don't recycle requests
-                        if (!requestId || !uniqueKey || this.inProgress.has(requestId) || this.recentlyHandled.get(requestId!)) return;
+                        if (!requestId || !uniqueKey || this.requestIdsInProgress.has(requestId) || this.recentlyHandledRequestsCache.get(requestId!)) return;
 
-                        this.queueHeadDict.add(requestId, requestId, false);
+                        this.queueHeadIds.add(requestId, requestId, false);
                         this._cacheRequest(getRequestId(uniqueKey), {
                             requestId,
                             wasAlreadyHandled: false,
@@ -740,7 +238,7 @@ export class RequestQueue {
         if (prevLimit >= REQUEST_QUEUE_HEAD_MAX_LIMIT) {
             this.log.warning(`Reached the maximum number of requests in progress: ${REQUEST_QUEUE_HEAD_MAX_LIMIT}.`);
         }
-        const shouldRepeatWithHigherLimit = this.queueHeadDict.length() === 0
+        const shouldRepeatWithHigherLimit = this.queueHeadIds.length() === 0
             && wasLimitReached
             && prevLimit < REQUEST_QUEUE_HEAD_MAX_LIMIT;
 
@@ -772,205 +270,4 @@ export class RequestQueue {
 
         return this._ensureHeadIsNonEmpty(ensureConsistency, nextLimit, iteration + 1);
     }
-
-    /**
-     * Adds a request straight to the queueHeadDict, to improve performance.
-     */
-    private _maybeAddRequestToQueueHead(requestId: string, forefront: boolean): void {
-        if (forefront) {
-            this.queueHeadDict.add(requestId, requestId, true);
-        } else if (this.assumedTotalCount < QUERY_HEAD_MIN_LENGTH) {
-            this.queueHeadDict.add(requestId, requestId, false);
-        }
-    }
-
-    /**
-     * Removes the queue either from the Apify Cloud storage or from the local database,
-     * depending on the mode of operation.
-     */
-    async drop(): Promise<void> {
-        await this.client.delete();
-        const manager = StorageManager.getManager(RequestQueue, this.config);
-        manager.closeStorage(this);
-    }
-
-    /**
-     * Returns the number of handled requests.
-     *
-     * This function is just a convenient shortcut for:
-     *
-     * ```javascript
-     * const { handledRequestCount } = await queue.getInfo();
-     * ```
-     */
-    async handledCount(): Promise<number> {
-        // NOTE: We keep this function for compatibility with RequestList.handledCount()
-        const { handledRequestCount } = await this.getInfo() ?? {};
-        return handledRequestCount ?? 0;
-    }
-
-    /**
-     * Returns an object containing general information about the request queue.
-     *
-     * The function returns the same object as the Apify API Client's
-     * [getQueue](https://docs.apify.com/api/apify-client-js/latest#ApifyClient-requestQueues)
-     * function, which in turn calls the
-     * [Get request queue](https://apify.com/docs/api/v2#/reference/request-queues/queue/get-request-queue)
-     * API endpoint.
-     *
-     * **Example:**
-     * ```
-     * {
-     *   id: "WkzbQMuFYuamGv3YF",
-     *   name: "my-queue",
-     *   userId: "wRsJZtadYvn4mBZmm",
-     *   createdAt: new Date("2015-12-12T07:34:14.202Z"),
-     *   modifiedAt: new Date("2015-12-13T08:36:13.202Z"),
-     *   accessedAt: new Date("2015-12-14T08:36:13.202Z"),
-     *   totalRequestCount: 25,
-     *   handledRequestCount: 5,
-     *   pendingRequestCount: 20,
-     * }
-     * ```
-     */
-    async getInfo(): Promise<RequestQueueInfo | undefined> {
-        return this.client.get();
-    }
-
-    /**
-     * Fetches URLs from requestsFromUrl and returns them in format of list of requests
-     */
-    protected async _fetchRequestsFromUrl(source: InternalSource): Promise<RequestOptions[]> {
-        const { requestsFromUrl, regex, ...sharedOpts } = source;
-
-        // Download remote resource and parse URLs.
-        let urlsArr;
-        try {
-            urlsArr = await this._downloadListOfUrls({ url: requestsFromUrl, urlRegExp: regex, proxyUrl: await this.proxyConfiguration?.newUrl() });
-        } catch (err) {
-            throw new Error(`Cannot fetch a request list from ${requestsFromUrl}: ${err}`);
-        }
-
-        // Skip if resource contained no URLs.
-        if (!urlsArr.length) {
-            this.log.warning('list fetched, but it is empty.', { requestsFromUrl, regex });
-            return [];
-        }
-
-        return urlsArr.map((url) => ({ url, ...sharedOpts }));
-    }
-
-    /**
-     * Adds all fetched requests from a URL from a remote resource.
-     */
-    protected async _addFetchedRequests(source: InternalSource, fetchedRequests: RequestOptions[], options: RequestQueueOperationOptions) {
-        const { requestsFromUrl, regex } = source;
-        const { addedRequests } = await this.addRequestsBatched(fetchedRequests, options);
-
-        this.log.info('Fetched and loaded Requests from a remote resource.', {
-            requestsFromUrl,
-            regex,
-            fetchedCount: fetchedRequests.length,
-            importedCount: addedRequests.length,
-            duplicateCount: fetchedRequests.length - addedRequests.length,
-            sample: JSON.stringify(fetchedRequests.slice(0, 5)),
-        });
-
-        return addedRequests;
-    }
-
-    /**
-     * @internal wraps public utility for mocking purposes
-     */
-    private async _downloadListOfUrls(options: { url: string; urlRegExp?: RegExp; proxyUrl?: string }): Promise<string[]> {
-        return downloadListOfUrls(options);
-    }
-
-    /**
-     * Opens a request queue and returns a promise resolving to an instance
-     * of the {@apilink RequestQueue} class.
-     *
-     * {@apilink RequestQueue} represents a queue of URLs to crawl, which is stored either on local filesystem or in the cloud.
-     * The queue is used for deep crawling of websites, where you start with several URLs and then
-     * recursively follow links to other pages. The data structure supports both breadth-first
-     * and depth-first crawling orders.
-     *
-     * For more details and code examples, see the {@apilink RequestQueue} class.
-     *
-     * @param [queueIdOrName]
-     *   ID or name of the request queue to be opened. If `null` or `undefined`,
-     *   the function returns the default request queue associated with the crawler run.
-     * @param [options] Open Request Queue options.
-     */
-    static async open(queueIdOrName?: string | null, options: StorageManagerOptions = {}): Promise<RequestQueue> {
-        ow(queueIdOrName, ow.optional.any(ow.string, ow.null));
-        ow(options, ow.object.exactShape({
-            config: ow.optional.object.instanceOf(Configuration),
-            storageClient: ow.optional.object,
-            proxyConfiguration: ow.optional.object,
-        }));
-
-        options.config ??= Configuration.getGlobalConfig();
-        options.storageClient ??= options.config.getStorageClient();
-
-        await purgeDefaultStorages({ onlyPurgeOnce: true, client: options.storageClient, config: options.config });
-
-        const manager = StorageManager.getManager(this, options.config);
-        const queue = await manager.openStorage(queueIdOrName, options.storageClient);
-        queue.proxyConfiguration = options.proxyConfiguration;
-
-        return queue;
-    }
-}
-
-export interface RequestQueueOptions {
-    id: string;
-    name?: string;
-    client: StorageClient;
-
-    /**
-     * Used to pass the proxy configuration for the `requestsFromUrl` objects.
-     * Takes advantage of the internal address rotation and authentication process.
-     * If undefined, the `requestsFromUrl` requests will be made without proxy.
-     */
-    proxyConfiguration?: ProxyConfiguration;
-}
-
-export interface AddRequestsBatchedOptions extends RequestQueueOperationOptions {
-    /**
-     * Whether to wait for all the provided requests to be added, instead of waiting just for the initial batch of up to `batchSize`.
-     * @default false
-     */
-    waitForAllRequestsToBeAdded?: boolean;
-
-    /**
-     * @default 1000
-     */
-    batchSize?: number;
-
-    /**
-     * @default 1000
-     */
-    waitBetweenBatchesMillis?: number;
-}
-
-export interface AddRequestsBatchedResult {
-    addedRequests: ProcessedRequest[];
-    /**
-     * A promise which will resolve with the rest of the requests that were added to the queue.
-     *
-     * Alternatively, we can set {@apilink AddRequestsBatchedOptions.waitForAllRequestsToBeAdded|`waitForAllRequestsToBeAdded`} to `true`
-     * in the {@apilink BasicCrawler.addRequests|`crawler.addRequests()`} options.
-     *
-     * **Example:**
-     *
-     * ```ts
-     * // Assuming `requests` is a list of requests.
-     * const result = await crawler.addRequests(requests);
-     *
-     * // If we want to wait for the rest of the requests to be added to the queue:
-     * await result.waitForAllRequestsToBeAdded;
-     * ```
-     */
-    waitForAllRequestsToBeAdded: Promise<ProcessedRequest[]>;
 }
