+++ packages/linkedom-crawler/src/internals/linkedom-crawler.ts
@@ -8,10 +8,9 @@ import type {
     ErrorHandler,
     RequestHandler,
     EnqueueLinksOptions,
-    RequestQueue,
     GetUserDataFromRequest,
     RouterRoutes,
-    RequestQueueV2,
+    RequestProvider,
 } from '@crawlee/http';
 import { HttpCrawler, enqueueLinks, Router, resolveBaseUrlForEnqueueLinksFiltering, tryAbsoluteURL } from '@crawlee/http';
 import type { Dictionary } from '@crawlee/types';
@@ -170,7 +169,7 @@ export class LinkeDOMCrawler extends HttpCrawler<LinkeDOMCrawlingContext> {
 interface EnqueueLinksInternalOptions {
     options?: LinkeDOMCrawlerEnqueueLinksOptions;
     window: Window | null;
-    requestQueue: RequestQueue | RequestQueueV2;
+    requestQueue: RequestProvider;
     originalRequestUrl: string;
     finalRequestUrl?: string;
 }
