+++ packages/playwright-crawler/src/internals/enqueue-links/click-elements.ts
@@ -7,9 +7,8 @@ import type {
     RegExpInput,
     RequestTransform,
     UrlPatternObject,
-    RequestQueue,
     RequestOptions,
-    RequestQueueV2,
+    RequestProvider,
 } from '@crawlee/browser';
 import {
     constructGlobObjectsFromGlobs,
@@ -41,7 +40,7 @@ export interface EnqueueLinksByClickingElementsOptions {
     /**
      * A request queue to which the URLs will be enqueued.
      */
-    requestQueue: RequestQueue | RequestQueueV2;
+    requestQueue: RequestProvider;
 
     /**
      * A CSS selector matching elements to be clicked on. Unlike in {@apilink enqueueLinks}, there is no default
